+++ Software/Audio2Led/Drivers/msgeq7.h
@@ -22,7 +22,7 @@
 
 #define MSGEQ7_AUD_BANDS   7       // Audio bands in MSGEQ7
 
-void msgeq7_get_audio(uint16_t *audio);
+void msgeq7_get_audio(uint8_t *audio);
 void msgeq7_init(void);
 
 #endif /* _MSGEQ7_H */
